Fix remember me on users

+++ db/migrate/076_add_login_to_users.rb
@@ -1,10 +1,15 @@
 class AddLoginToUsers < ActiveRecord::Migration
   def self.up
-    add_column :users, :login, :string
-    add_column :users, :salt, :string, :limit => 40
+    # only migrate systems that don't tie into the opencongress database
+    unless ActiveRecord::Base.configurations.has_key?('opencongress')
+      add_column :users, :login, :string
+      add_column :users, :salt, :string, :limit => 40
+    end
   end
 
   def self.down
-    remove_column :users, :login, :salt
+    unless ActiveRecord::Base.configurations.has_key?('opencongress')
+      remove_column :users, :login, :salt
+    end
   end
 end
